refactor: split large files, extract duplicates, remove dead shims
Aggressive refactoring of the features layer, consensus-approved via
RALPLAN-DR (Architect + Critic, 2 iterations).

Phase 1 — Extract shared helpers:
- _codegen_snippets.py: geometry/alignment/classify code-string preambles
- _subprocess.py: TotalSegmentator subprocess management (breaks
  diagnostics/ → spine/ cross-domain coupling)
- _validation.py: shared _validate_region across ct, onco_spine, modic
- Migrate VALID_* constants from spine/tools.py to spine/constants.py

Phase 2 — Standardize code generation:
- core/codegen.py: safe_string, safe_json, safe_optional helpers
- Boolean audit: confirmed zero json.dumps(bool) in features/

Phase 3 — Split spine/tools.py (2424 LOC → 4 modules):
- spine/ccj.py (419 LOC): CCJ landmark extraction + angle measurements
- spine/alignment.py (455 LOC): sagittal alignment parameters
- spine/segmentation.py (753 LOC): TotalSegmentator spine + VA segmentation
- spine/analysis.py (577 LOC): bone quality + clinical visualization
- spine/tools.py: 55-line re-export hub

Phase 4 — Split diagnostics/ct.py (1901 LOC → 3 modules):
- ct_fracture.py (876 LOC): fracture, listhesis, canal morphometry
- ct_metabolic.py (591 LOC): osteoporosis, metastatic lesions
- ct_scoring.py (454 LOC): SINS score calculation
- diagnostics/_common.py (35 LOC): shared _validate_levels
- ct.py: 70-line re-export hub

Phase 5 — Remove 13 dead backward-compat shim files (~291 LOC removed).
Keep spine_constants.py (active test consumer).

Coverage: 94.39% (was 93%). Tests: 1402 passed, 0 failed.
