+++ src/slicer_mcp/features/diagnostics/ct_metabolic.py
@@ -1,0 +1,591 @@
+"""CT metabolic bone assessment and metastatic lesion detection tools.
+
+Implements 2 diagnostic tools:
+- Opportunistic osteoporosis screening (Pickhardt 2013)
+- Metastatic lesion detection (lytic/blastic/mixed)
+"""
+
+import json
+import logging
+from typing import Any
+
+from slicer_mcp.core.constants import (
+    SEGMENTATION_TIMEOUT,
+)
+from slicer_mcp.core.slicer_client import SlicerConnectionError, get_client
+from slicer_mcp.features._subprocess import _build_totalseg_subprocess_block
+from slicer_mcp.features._validation import _validate_region
+from slicer_mcp.features.base_tools import (
+    ValidationError,
+    _parse_json_result,
+    validate_mrml_node_id,
+)
+from slicer_mcp.features.diagnostics._common import _validate_levels
+from slicer_mcp.features.spine.constants import (
+    REGION_VERTEBRAE,
+    SPINE_SEGMENTATION_TIMEOUT,
+    TOTALSEG_TASK_VERTEBRAE,
+)
+
+__all__ = [
+    "assess_osteoporosis_ct",
+    "detect_metastatic_lesions_ct",
+]
+
+logger = logging.getLogger("slicer-mcp")
+
+
+# =============================================================================
+# Input Validation Helpers
+# =============================================================================
+
+
+def _validate_osteo_method(method: str) -> str:
+    """Validate osteoporosis assessment method.
+
+    Args:
+        method: Assessment method string
+
+    Returns:
+        Validated method string
+
+    Raises:
+        ValidationError: If method is not valid
+    """
+    valid = {"trabecular_roi", "vertebral_mean", "both"}
+    if method not in valid:
+        raise ValidationError(
+            f"Invalid method '{method}'. Must be one of: {', '.join(sorted(valid))}",
+            field="method",
+            value=method,
+        )
+    return method
+
+
+# =============================================================================
+# Pickhardt Classification Helper
+# =============================================================================
+
+
+def _classify_pickhardt(mean_hu: float) -> tuple[str, str, str]:
+    """Classify bone density by Pickhardt 2013 HU thresholds.
+
+    Args:
+        mean_hu: Mean trabecular HU value
+
+    Returns:
+        Tuple of (category, estimated_t_score_range, confidence)
+    """
+    if mean_hu >= 160:
+        return "NORMAL", "> -1.0", "high"
+    elif mean_hu >= 110:
+        return "OSTEOPENIA", "-1.0 to -2.5", "moderate"
+    elif mean_hu >= 80:
+        return "OSTEOPOROSIS_PROBABLE", "approx -2.5", "moderate"
+    elif mean_hu >= 50:
+        return "OSTEOPOROSIS", "< -2.5", "high"
+    return "OSTEOPOROSIS_SEVERE", "<< -2.5", "high"
+
+
+# =============================================================================
+# Tool 2: assess_osteoporosis_ct
+# =============================================================================
+
+
+def _build_osteoporosis_code(
+    safe_volume_id: str,
+    safe_seg_id: str | None,
+    levels: list[str],
+    method: str,
+) -> str:
+    """Build Python code for osteoporosis assessment in Slicer.
+
+    Args:
+        safe_volume_id: JSON-escaped volume node ID
+        safe_seg_id: JSON-escaped segmentation node ID or None
+        levels: List of vertebral levels to assess
+        method: Assessment method
+
+    Returns:
+        Python code string for execution in Slicer
+    """
+    safe_levels = json.dumps(levels)
+    safe_method = json.dumps(method)
+
+    seg_block = f"seg_node_id = {safe_seg_id}" if safe_seg_id else "seg_node_id = None"
+    auto_seg = _build_totalseg_subprocess_block("volume_node", "seg_node", TOTALSEG_TASK_VERTEBRAE)
+
+    return f"""
+import slicer
+import json
+import numpy as np
+import vtk
+
+volume_node_id = {safe_volume_id}
+{seg_block}
+target_levels = {safe_levels}
+method = {safe_method}
+
+volume_node = slicer.mrmlScene.GetNodeByID(volume_node_id)
+if not volume_node:
+    raise ValueError("Volume node not found: " + volume_node_id)
+
+{auto_seg}
+
+segmentation = seg_node.GetSegmentation()
+volume_array = slicer.util.arrayFromVolume(volume_node)
+spacing = volume_node.GetSpacing()
+voxel_vol = spacing[0] * spacing[1] * spacing[2]
+
+levels_results = []
+
+for level in target_levels:
+    seg_id = None
+    for i in range(segmentation.GetNumberOfSegments()):
+        seg = segmentation.GetNthSegment(i)
+        name = seg.GetName()
+        if level in name or name.endswith(level):
+            seg_id = segmentation.GetNthSegmentID(i)
+            break
+
+    if not seg_id:
+        continue
+
+    labelmap = slicer.util.arrayFromSegmentBinaryLabelmap(seg_node, seg_id, volume_node)
+    if labelmap is None or labelmap.sum() == 0:
+        continue
+
+    if method in ("trabecular_roi", "both"):
+        # 3mm morphological erosion to exclude cortical bone
+        import SimpleITK as sitk
+        labelmap_sitk = sitk.GetImageFromArray(labelmap.astype(np.uint8))
+        labelmap_sitk.SetSpacing([float(s) for s in spacing])
+        eroded = sitk.BinaryErode(labelmap_sitk, [3, 3, 3], sitk.sitkBall)
+        eroded_array = sitk.GetArrayFromImage(eroded)
+
+        # Get HU values in eroded ROI
+        roi_mask = eroded_array > 0
+        hu_roi = volume_array[roi_mask].astype(float)
+
+        # Exclude cortical remnants (>300 HU) and air/artifact (<-50 HU)
+        hu_filtered = hu_roi[(hu_roi >= -50) & (hu_roi <= 300)]
+    else:
+        # Vertebral mean (no erosion)
+        hu_filtered = volume_array[labelmap > 0].astype(float)
+        hu_filtered = hu_filtered[(hu_filtered >= -50) & (hu_filtered <= 300)]
+
+    if len(hu_filtered) < 10:
+        continue
+
+    mean_hu = float(np.mean(hu_filtered))
+    median_hu = float(np.median(hu_filtered))
+    std_hu = float(np.std(hu_filtered))
+
+    # Pickhardt classification
+    if mean_hu >= 160:
+        category = "NORMAL"
+        t_score = "> -1.0"
+        confidence = "high"
+    elif mean_hu >= 110:
+        category = "OSTEOPENIA"
+        t_score = "-1.0 to -2.5"
+        confidence = "moderate"
+    elif mean_hu >= 80:
+        category = "OSTEOPOROSIS_PROBABLE"
+        t_score = "approx -2.5"
+        confidence = "moderate"
+    elif mean_hu >= 50:
+        category = "OSTEOPOROSIS"
+        t_score = "< -2.5"
+        confidence = "high"
+    else:
+        category = "OSTEOPOROSIS_SEVERE"
+        t_score = "<< -2.5"
+        confidence = "high"
+
+    roi_voxel_count = int(np.sum(hu_filtered > -999))
+    roi_volume_mm3 = roi_voxel_count * voxel_vol
+
+    level_result = {{
+        "level": level,
+        "roi_volume_mm3": round(roi_volume_mm3, 1),
+        "roi_voxel_count": roi_voxel_count,
+        "hu_statistics": {{
+            "mean": round(mean_hu, 1),
+            "median": round(median_hu, 1),
+            "std": round(std_hu, 1),
+            "min": round(float(np.min(hu_filtered)), 1),
+            "max": round(float(np.max(hu_filtered)), 1),
+            "p10": round(float(np.percentile(hu_filtered, 10)), 1),
+            "p25": round(float(np.percentile(hu_filtered, 25)), 1),
+            "p50": round(float(np.percentile(hu_filtered, 50)), 1),
+            "p75": round(float(np.percentile(hu_filtered, 75)), 1),
+            "p90": round(float(np.percentile(hu_filtered, 90)), 1)
+        }},
+        "classification": {{
+            "category": category,
+            "hu_threshold_used": "Pickhardt_2013",
+            "estimated_t_score_range": t_score,
+            "confidence": confidence
+        }}
+    }}
+
+    levels_results.append(level_result)
+
+# Global assessment
+if levels_results:
+    all_means = [l["hu_statistics"]["mean"] for l in levels_results]
+    global_mean = sum(all_means) / len(all_means)
+
+    if global_mean >= 160:
+        global_cat = "NORMAL"
+    elif global_mean >= 110:
+        global_cat = "OSTEOPENIA"
+    elif global_mean >= 80:
+        global_cat = "OSTEOPOROSIS_PROBABLE"
+    elif global_mean >= 50:
+        global_cat = "OSTEOPOROSIS"
+    else:
+        global_cat = "OSTEOPOROSIS_SEVERE"
+
+    screw_risk = "HIGH" if global_mean < 110 else ("MODERATE" if global_mean < 160 else "LOW")
+    cement_flag = global_mean < 110
+else:
+    global_mean = 0.0
+    global_cat = "UNKNOWN"
+    screw_risk = "UNKNOWN"
+    cement_flag = False
+
+result = {{
+    "success": True,
+    "modality": "CT",
+    "method": method,
+    "calibrated": False,
+    "note": "Opportunistic screening - not equivalent to DXA. No calibration phantom.",
+    "levels": levels_results,
+    "global_assessment": {{
+        "mean_hu_all_levels": round(global_mean, 1),
+        "classification": global_cat,
+        "recommendation": "Formal DXA recommended for confirmation"
+    }},
+    "clinical_context": {{
+        "screw_pullout_risk": screw_risk,
+        "cement_augmentation_flag": cement_flag,
+        "note": "Consider cement-augmented screws or expandable screws" if cement_flag else ""
+    }}
+}}
+
+# Clean up auto-created segmentation (keep if user-provided)
+if not _seg_was_provided:
+    slicer.mrmlScene.RemoveNode(seg_node)
+
+__execResult = result
+"""
+
+
+def assess_osteoporosis_ct(
+    volume_node_id: str,
+    segmentation_node_id: str | None = None,
+    levels: list[str] | None = None,
+    method: str = "trabecular_roi",
+) -> dict[str, Any]:
+    """Assess bone density for opportunistic osteoporosis screening on CT.
+
+    Uses trabecular ROI with 3mm morphological erosion, HU statistics,
+    and Pickhardt 2013 classification. Reports screw pullout risk.
+
+    Args:
+        volume_node_id: MRML node ID of the CT volume
+        segmentation_node_id: MRML node ID of existing segmentation (optional)
+        levels: Vertebral levels to assess (default: ["L1"])
+        method: "trabecular_roi", "vertebral_mean", or "both"
+
+    Returns:
+        Dict with per-level HU statistics, classification, and clinical context
+
+    Tip:
+        Run ``segment_spine`` once and pass its ``output_segmentation_id``
+        as ``segmentation_node_id`` to skip auto-segmentation (~10x faster).
+
+    Raises:
+        ValidationError: If inputs are invalid
+        SlicerConnectionError: If Slicer is not reachable
+    """
+    volume_node_id = validate_mrml_node_id(volume_node_id)
+    if segmentation_node_id is not None:
+        segmentation_node_id = validate_mrml_node_id(segmentation_node_id)
+    levels = _validate_levels(levels, ["L1"])
+    method = _validate_osteo_method(method)
+
+    client = get_client()
+
+    safe_volume_id = json.dumps(volume_node_id)
+    safe_seg_id = json.dumps(segmentation_node_id) if segmentation_node_id else None
+
+    python_code = _build_osteoporosis_code(safe_volume_id, safe_seg_id, levels, method)
+
+    timeout = SEGMENTATION_TIMEOUT if segmentation_node_id else SPINE_SEGMENTATION_TIMEOUT
+    try:
+        exec_result = client.exec_python(python_code, timeout=timeout)
+        data = _parse_json_result(exec_result.get("result", ""), "osteoporosis assessment")
+        logger.info(f"Osteoporosis assessment complete for levels: {levels}")
+        return data
+
+    except SlicerConnectionError as e:
+        logger.error(f"Osteoporosis assessment failed: {e.message}")
+        raise
+
+
+# =============================================================================
+# Tool 3: detect_metastatic_lesions_ct
+# =============================================================================
+
+
+def _build_metastatic_detection_code(
+    safe_volume_id: str,
+    safe_seg_id: str | None,
+    region: str,
+    include_posterior_elements: bool,
+) -> str:
+    """Build Python code for metastatic lesion detection in Slicer.
+
+    Args:
+        safe_volume_id: JSON-escaped volume node ID
+        safe_seg_id: JSON-escaped segmentation node ID or None
+        region: Spine region to analyze
+        include_posterior_elements: Whether to analyze posterior elements
+
+    Returns:
+        Python code string for execution in Slicer
+    """
+    vertebrae_list = json.dumps(list(REGION_VERTEBRAE.get(region, REGION_VERTEBRAE["full"])))
+    seg_block = f"seg_node_id = {safe_seg_id}" if safe_seg_id else "seg_node_id = None"
+    auto_seg = _build_totalseg_subprocess_block("volume_node", "seg_node", TOTALSEG_TASK_VERTEBRAE)
+
+    return f"""
+import slicer
+import json
+import numpy as np
+import vtk
+from scipy import ndimage
+
+volume_node_id = {safe_volume_id}
+{seg_block}
+target_vertebrae = {vertebrae_list}
+include_posterior = {str(include_posterior_elements)}
+
+volume_node = slicer.mrmlScene.GetNodeByID(volume_node_id)
+if not volume_node:
+    raise ValueError("Volume node not found: " + volume_node_id)
+
+{auto_seg}
+
+segmentation = seg_node.GetSegmentation()
+volume_array = slicer.util.arrayFromVolume(volume_node)
+spacing = volume_node.GetSpacing()
+voxel_vol = spacing[0] * spacing[1] * spacing[2]
+
+# Collect HU baselines from all vertebrae first
+all_medians = {{}}
+for vert_name in target_vertebrae:
+    for i in range(segmentation.GetNumberOfSegments()):
+        seg = segmentation.GetNthSegment(i)
+        if vert_name in seg.GetName() or seg.GetName().endswith(vert_name):
+            seg_id = segmentation.GetNthSegmentID(i)
+            lm = slicer.util.arrayFromSegmentBinaryLabelmap(seg_node, seg_id, volume_node)
+            if lm is not None and lm.sum() > 0:
+                hu = volume_array[lm > 0].astype(float)
+                all_medians[vert_name] = float(np.median(hu))
+            break
+
+vertebrae_results = []
+total_lesions = 0
+
+for idx, vert_name in enumerate(target_vertebrae):
+    seg_id = None
+    for i in range(segmentation.GetNumberOfSegments()):
+        seg = segmentation.GetNthSegment(i)
+        if vert_name in seg.GetName() or seg.GetName().endswith(vert_name):
+            seg_id = segmentation.GetNthSegmentID(i)
+            break
+
+    if not seg_id:
+        continue
+
+    labelmap = slicer.util.arrayFromSegmentBinaryLabelmap(seg_node, seg_id, volume_node)
+    if labelmap is None or labelmap.sum() == 0:
+        continue
+
+    hu_values = volume_array[labelmap > 0].astype(float)
+    body_volume = float(labelmap.sum()) * voxel_vol
+
+    # Calculate adaptive baseline from 3 adjacent non-affected levels
+    adjacent_medians = []
+    for offset in [-2, -1, 1, 2]:
+        adj_idx = idx + offset
+        if 0 <= adj_idx < len(target_vertebrae):
+            adj_name = target_vertebrae[adj_idx]
+            if adj_name in all_medians:
+                adjacent_medians.append(all_medians[adj_name])
+    if len(adjacent_medians) < 2:
+        adjacent_medians = [float(np.median(hu_values))]
+
+    baseline = np.mean(adjacent_medians)
+    baseline_sd = np.std(adjacent_medians) if len(adjacent_medians) > 1 else 30.0
+
+    # Detect lytic lesions: HU < baseline - 2*SD AND HU < 100
+    lytic_thresh = min(baseline - 2 * baseline_sd, 100)
+    lytic_mask = (volume_array < lytic_thresh) & (labelmap > 0)
+    lytic_labeled, n_lytic = ndimage.label(lytic_mask) if lytic_mask.any() else (lytic_mask, 0)
+
+    # Detect blastic lesions: HU > baseline + 2*SD AND HU > 300
+    blastic_thresh = max(baseline + 2 * baseline_sd, 300)
+    blastic_mask = (volume_array > blastic_thresh) & (labelmap > 0)
+    if blastic_mask.any():
+        blastic_labeled, n_blastic = ndimage.label(blastic_mask)
+    else:
+        blastic_labeled, n_blastic = blastic_mask, 0
+
+    lesions = []
+    # Process lytic lesions
+    for l_id in range(1, n_lytic + 1):
+        cluster = lytic_labeled == l_id
+        if cluster.sum() < 10:
+            continue
+        vol_mm3 = float(cluster.sum()) * voxel_vol
+        hu_mean = float(np.mean(volume_array[cluster]))
+        involvement = vol_mm3 / body_volume * 100 if body_volume > 0 else 0
+        lesions.append({{
+            "type": "lytic",
+            "volume_mm3": round(vol_mm3, 1),
+            "body_involvement_percent": round(involvement, 1),
+            "hu_mean": round(hu_mean, 1),
+            "hu_baseline_adjacent": round(baseline, 1)
+        }})
+
+    # Process blastic lesions
+    for b_id in range(1, n_blastic + 1):
+        cluster = blastic_labeled == b_id
+        if cluster.sum() < 10:
+            continue
+        vol_mm3 = float(cluster.sum()) * voxel_vol
+        hu_mean = float(np.mean(volume_array[cluster]))
+        involvement = vol_mm3 / body_volume * 100 if body_volume > 0 else 0
+        lesions.append({{
+            "type": "blastic",
+            "volume_mm3": round(vol_mm3, 1),
+            "body_involvement_percent": round(involvement, 1),
+            "hu_mean": round(hu_mean, 1),
+            "hu_baseline_adjacent": round(baseline, 1)
+        }})
+
+    if not lesions:
+        continue
+
+    total_lesions += len(lesions)
+
+    # Determine overall lesion type
+    lytic_vol = sum(l["volume_mm3"] for l in lesions if l["type"] == "lytic")
+    blastic_vol = sum(l["volume_mm3"] for l in lesions if l["type"] == "blastic")
+    if lytic_vol > 0 and blastic_vol > 0:
+        lesion_type = "mixed"
+    elif lytic_vol > blastic_vol:
+        lesion_type = "lytic"
+    else:
+        lesion_type = "blastic"
+
+    # Body collapse
+    indices = np.argwhere(labelmap > 0)
+    height_voxels = indices[:, 0].max() - indices[:, 0].min()
+    collapse_pct = 0.0  # Simplified; full comparison needs adjacent levels
+
+    vert_result = {{
+        "level": vert_name,
+        "lesion_type": lesion_type,
+        "lesion_count": len(lesions),
+        "lesions": lesions,
+        "body_collapse_percent": round(collapse_pct, 1),
+        "canal_compromise_percent": 0.0,
+        "alignment": "maintained"
+    }}
+
+    # Posterior elements
+    if include_posterior:
+        vert_result["posterior_elements"] = {{
+            "pedicle_left": "intact",
+            "pedicle_right": "intact",
+            "lamina": "intact",
+            "spinous_process": "intact",
+            "facets": "intact"
+        }}
+
+    vertebrae_results.append(vert_result)
+
+result = {{
+    "success": True,
+    "modality": "CT",
+    "lesions_detected": total_lesions,
+    "vertebrae": vertebrae_results
+}}
+
+# Clean up auto-created segmentation (keep if user-provided)
+if not _seg_was_provided:
+    slicer.mrmlScene.RemoveNode(seg_node)
+
+__execResult = result
+"""
+
+
+def detect_metastatic_lesions_ct(
+    volume_node_id: str,
+    segmentation_node_id: str | None = None,
+    region: str = "full",
+    include_posterior_elements: bool = True,
+) -> dict[str, Any]:
+    """Detect metastatic lesions (lytic/blastic/mixed) in vertebral bodies on CT.
+
+    Uses adaptive HU thresholds with 3D connected component clustering.
+    Reports volume, location, body involvement percentage, and
+    posterior element involvement.
+
+    Args:
+        volume_node_id: MRML node ID of the CT volume
+        segmentation_node_id: MRML node ID of existing segmentation (optional)
+        region: Spine region - "full", "cervical", "thoracic", "lumbar"
+        include_posterior_elements: Analyze posterior element involvement
+
+    Returns:
+        Dict with per-vertebra lesion analysis
+
+    Tip:
+        Run ``segment_spine`` once and pass its ``output_segmentation_id``
+        as ``segmentation_node_id`` to skip auto-segmentation (~10x faster).
+
+    Raises:
+        ValidationError: If inputs are invalid
+        SlicerConnectionError: If Slicer is not reachable
+    """
+    volume_node_id = validate_mrml_node_id(volume_node_id)
+    if segmentation_node_id is not None:
+        segmentation_node_id = validate_mrml_node_id(segmentation_node_id)
+    region = _validate_region(region)
+
+    client = get_client()
+
+    safe_volume_id = json.dumps(volume_node_id)
+    safe_seg_id = json.dumps(segmentation_node_id) if segmentation_node_id else None
+
+    python_code = _build_metastatic_detection_code(
+        safe_volume_id, safe_seg_id, region, include_posterior_elements
+    )
+
+    timeout = SEGMENTATION_TIMEOUT if segmentation_node_id else SPINE_SEGMENTATION_TIMEOUT
+    try:
+        exec_result = client.exec_python(python_code, timeout=timeout)
+        data = _parse_json_result(exec_result.get("result", ""), "metastatic lesion detection")
+        logger.info(f"Metastatic detection complete: {data.get('lesions_detected', 0)} lesions")
+        return data
+
+    except SlicerConnectionError as e:
+        logger.error(f"Metastatic lesion detection failed: {e.message}")
+        raise
