+++ tests/unit/test_shim_backward_compat.py
@@ -1,7 +1,8 @@
-"""Tests that the 14 root-level shim modules correctly re-export canonical symbols.
+"""Tests that the remaining shim module correctly re-exports canonical symbols.
 
-Each shim exists for backward compatibility and should re-export at least one
-key symbol from its canonical module.
+Only ``spine_constants`` remains as a backward-compat shim (active consumer
+in ``test_diagnostic_tools_ct.py``).  The other 13 shims were removed in the
+2026-04 aggressive refactoring.
 """
 
 import importlib
@@ -10,36 +11,7 @@ import warnings
 import pytest
 
 SHIM_TO_CANONICAL = [
-    ("slicer_mcp.circuit_breaker", "slicer_mcp.core.circuit_breaker", "CircuitBreaker"),
-    ("slicer_mcp.constants", "slicer_mcp.core.constants", "DEFAULT_SLICER_URL"),
-    ("slicer_mcp.metrics", "slicer_mcp.core.metrics", "track_request"),
-    ("slicer_mcp.resources", "slicer_mcp.core.resources", "get_status_resource"),
-    ("slicer_mcp.slicer_client", "slicer_mcp.core.slicer_client", "SlicerClient"),
-    ("slicer_mcp.tools", "slicer_mcp.features.base_tools", "capture_screenshot"),
-    (
-        "slicer_mcp.diagnostic_tools_ct",
-        "slicer_mcp.features.diagnostics.ct",
-        "detect_vertebral_fractures_ct",
-    ),
-    (
-        "slicer_mcp.diagnostic_tools_mri",
-        "slicer_mcp.features.diagnostics.mri",
-        "classify_modic_changes",
-    ),
-    (
-        "slicer_mcp.diagnostic_tools_xray",
-        "slicer_mcp.features.diagnostics.xray",
-        "measure_cobb_angle_xray",
-    ),
-    (
-        "slicer_mcp.instrumentation_tools",
-        "slicer_mcp.features.spine.instrumentation",
-        "plan_cervical_screws",
-    ),
-    ("slicer_mcp.registration_tools", "slicer_mcp.features.registration", "place_landmarks"),
-    ("slicer_mcp.rendering_tools", "slicer_mcp.features.rendering", "enable_volume_rendering"),
     ("slicer_mcp.spine_constants", "slicer_mcp.features.spine.constants", "SPINE_REGIONS"),
-    ("slicer_mcp.spine_tools", "slicer_mcp.features.spine.tools", "segment_spine"),
 ]
 
 
@@ -71,7 +43,6 @@ def test_shim_re_exports_canonical_symbol(shim_path, canonical_path, symbol):
 )
 def test_shim_emits_deprecation_warning(shim_path):
     """Importing a shim module emits a DeprecationWarning."""
-    # Force re-import by removing from cache
     import sys
 
     sys.modules.pop(shim_path, None)
